Created common consumer to dashboard file to prevent code duplication
