+++ kafka/consumer-service/consumer-service.js
@@ -1,22 +1,9 @@
 import { Kafka } from "kafkajs";
-import http from "http";
-import { 
-  promiseHandler as handler
-} from "./shared/utils.js";
-import {
-  aggregations,
-  saveMessage
-} from "./shared/aggregations.js";
+import { promiseHandler as handler } from "./shared/utils.js";
+import { saveMessage } from "./shared/aggregations.js";
+import { startExperiment } from "./shared/consumer-to-dashboard.js";
 
 const TOPIC_NAME = "air-quality-observation-topic";
-const EXPERIMENT_TIME_MS = (process.env.NUMBER_OF_MINUTES || 10) * 60 * 1000;
-const AGGREGATE_PUBLISH_RATE = process.env.AGGREGATE_PUBLISH_RATE || 1000;
-
-const commonRequestProperties = {
-  hostname: "dashboard-app",
-  port: 3000,
-  method: "POST",
-};
 
 const kafka = new Kafka({
   clientId: "consumer-service-1",
@@ -45,42 +32,9 @@ const consumer = kafka.consumer({
     console.log(`Successfully subscribed to topic ${TOPIC_NAME}!`);
   }
 
-  setInterval(() => {
-    const data = JSON.stringify(aggregations);
-    const request = http.request({
-      path: "/aggregations",
-      headers: {
-        "Content-Type": "application/json",
-        "Content-Length": data.length
-      },
-      ...commonRequestProperties
-    });
-    request.write(data);
-    request.end();
+  // Start publish interval, experiment timeout and inform dashboard-backend
+  startExperiment();
 
-    // Empty aggregations after every time we send to dashboard backend
-    Object.keys(aggregations).forEach(stationId => delete aggregations[stationId]);
-
-  }, AGGREGATE_PUBLISH_RATE);
-
-  // A request to the dashboard backend for when the consumer will initialize its timeout function
-  // Used to make it possible to have a decently correct countdown on the frontend
-  {
-    const request = http.request({
-      path: "/start",
-      ...commonRequestProperties
-    });
-    request.end();
-  }
-
-  setTimeout(() => {
-    const request = http.request({
-      path: "/completed",
-      ...commonRequestProperties
-    });
-    request.end();
-  }, EXPERIMENT_TIME_MS);
-  
   // Run consumer and handle one message at a time
   await handler(consumer.run({
     eachMessage: async ({ topic, partition, message }) => {
